requirements.txt
pandas
requests
pymongo
python-dotenv
numpy
scikit-learn
